requirements.txt
streamlit
python-dotenv
langchain
langchain_community
PyPDF2
pinecone
faiss-cpu
langchain_google_genai
google-generativeai